// src/lib/gen/glimway/v1/calendar_pb.d.ts
// @generated by protoc-gen-es v2.12.0 with parameter "target=js+dts,import_extension=js"
// @generated from file glimway/v1/calendar.proto (package glimway.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file glimway/v1/calendar.proto.
 */
export declare const file_glimway_v1_calendar: GenFile;

/**
 * HTTP retains numeric Unix seconds, including dates beyond int32's 2038 limit.
 *
 * @generated from message glimway.v1.CalendarResponse
 */
export declare type CalendarResponse = Message<"glimway.v1.CalendarResponse"> & {
  /**
   * @generated from field: string wick = 1;
   */
  wick: string;

  /**
   * @generated from field: double wick_number = 2;
   */
  wickNumber: number;

  /**
   * @generated from field: double year = 3;
   */
  year: number;

  /**
   * @generated from field: int32 day = 4;
   */
  day: number;

  /**
   * @generated from field: string mark = 5;
   */
  mark: string;

  /**
   * @generated from field: google.protobuf.StringValue festival = 6;
   */
  festival?: string | undefined;

  /**
   * @generated from field: double starts_at = 7;
   */
  startsAt: number;

  /**
   * @generated from field: double next_turning = 8;
   */
  nextTurning: number;

  /**
   * @generated from field: google.protobuf.StringValue notice = 9;
   */
  notice?: string | undefined;

  /**
   * @generated from field: int32 wick_days = 10;
   */
  wickDays: number;
};

/**
 * Describes the message glimway.v1.CalendarResponse.
 * Use `create(CalendarResponseSchema)` to create a new message.
 */
export declare const CalendarResponseSchema: GenMessage<CalendarResponse>;